# shared/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AKIM AI — Аким на 5 часов",
    "description": "Учебная симуляция на синтетических данных. Детерминированные расчёты; AI объясняет готовые факты.",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/simulate": {
      "post": {
        "tags": [
          "Simulation"
        ],
        "summary": "Preview",
        "operationId": "preview_api_simulate_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScenarioRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Simulation"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/scenario/finalize": {
      "post": {
        "tags": [
          "Simulation"
        ],
        "summary": "Finalize",
        "operationId": "finalize_api_scenario_finalize_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScenarioRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Simulation"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/recommend": {
      "post": {
        "tags": [
          "Simulation"
        ],
        "summary": "Recommend One",
        "operationId": "recommend_one_api_recommend_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScenarioRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Recommendation"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Entity",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Budget": {
        "properties": {
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "spent": {
            "type": "integer",
            "title": "Spent"
          },
          "remaining": {
            "type": "integer",
            "title": "Remaining"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "total",
          "spent",
          "remaining"
        ],
        "title": "Budget"
      },
...
      "Contribution": {
        "properties": {
          "measureId": {
            "type": "string",
            "title": "Measureid"
          },
          "districtId": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "esil",
                  "almaty",
                  "saryarka",
                  "baikonur",
                  "nura"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Districtid"
          },
          "scoreImpact": {
            "type": "number",
            "title": "Scoreimpact"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "measureId",
          "scoreImpact"
        ],
        "title": "Contribution"
      },
      "CriticalIndicator": {
        "properties": {
          "districtId": {
            "type": "string",
            "enum": [
              "esil",
              "almaty",
              "saryarka",
              "baikonur",
              "nura"
            ],
            "title": "Districtid"
          },
          "indicatorId": {
            "type": "string",
            "enum": [
              "T1",
              "T2",
              "E1",
              "E2",
              "S1",
              "S2",
              "B1",
              "B2",
              "C1",
              "C2"
            ],
            "title": "Indicatorid"
          },
          "before": {
            "type": "number",
            "title": "Before"
          },
          "after": {
            "type": "number",
            "title": "After"
          },
          "resolved": {
            "type": "boolean",
            "title": "Resolved"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "districtId",
          "indicatorId",
          "before",
          "after",
          "resolved"
        ],
        "title": "CriticalIndicator"
      },
      "Decision": {
        "properties": {
          "measureId": {
            "type": "string",
            "maxLength": 64,
            "minLength": 1,
            "title": "Measureid"
          },
          "districtId": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 64,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Districtid"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "measureId"
        ],
        "title": "Decision"
      },
      "District": {
        "properties": {
          "id": {
            "type": "string",
            "enum": [
              "esil",
              "almaty",
              "saryarka",
              "baikonur",
              "nura"
            ],
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "populationShare": {
            "type": "number",
            "title": "Populationshare"
          },
          "profile": {
            "type": "string",
            "title": "Profile"
          },
          "indicators": {
            "additionalProperties": {
              "type": "number"
            },
            "propertyNames": {
              "enum": [
                "T1",
                "T2",
                "E1",
                "E2",
                "S1",
                "S2",
                "B1",
                "B2",
                "C1",
                "C2"
              ]
            },
            "type": "object",
            "title": "Indicators"
          },
          "score": {
            "type": "number",
            "title": "Score"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "name",
          "populationShare",
          "profile",
          "indicators",
          "score"
        ],
        "title": "District"
      },
      "ErrorResponse": {
        "properties": {
          "message": {
            "type": "string",
            "title": "Message"
          },
          "code": {
            "type": "string",
            "title": "Code"
          },
          "errors": {
            "items": {
              "$ref": "#/components/schemas/Issue"
            },
            "type": "array",
            "title": "Errors"
          },
          "score": {
            "type": "null",
            "title": "Score"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "message",
          "code"
        ],
        "title": "ErrorResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "Issue": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "measureIds": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Measureids"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "title": "Issue"
      },
...
      "Recommendation": {
        "properties": {
          "found": {
            "type": "boolean",
            "title": "Found"
          },
          "explanation": {
            "type": "string",
            "title": "Explanation"
          },
          "decisions": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/Decision"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Decisions"
          },
          "result": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Simulation"
              },
              {
                "type": "null"
              }
            ]
          },
          "candidatesChecked": {
            "type": "integer",
            "title": "Candidateschecked"
          },
          "validCandidates": {
            "type": "integer",
            "title": "Validcandidates"
          },
          "costDelta": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Costdelta"
          },
          "scoreDelta": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Scoredelta"
          },
          "weakestScoreDelta": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Weakestscoredelta"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "found",
          "explanation",
          "candidatesChecked",
          "validCandidates"
        ],
        "title": "Recommendation"
      },
...
      "ScenarioRequest": {
        "properties": {
          "decisions": {
            "items": {
              "$ref": "#/components/schemas/Decision"
            },
            "type": "array",
            "maxItems": 100,
            "title": "Decisions"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "decisions"
        ],
        "title": "ScenarioRequest"
      },
      "Simulation": {
        "properties": {
          "modelVersion": {
            "type": "string",
            "title": "Modelversion"
          },
          "dataChecksum": {
            "type": "string",
            "title": "Datachecksum"
          },
          "scenarioId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Scenarioid"
          },
          "source": {
            "type": "string",
            "const": "backend",
            "title": "Source",
            "default": "backend"
          },
          "decisions": {
            "items": {
              "$ref": "#/components/schemas/Decision"
            },
            "type": "array",
            "title": "Decisions"
          },
          "validation": {
            "$ref": "#/components/schemas/Validation"
          },
          "budget": {
            "$ref": "#/components/schemas/Budget"
          },
          "before": {
            "$ref": "#/components/schemas/Snapshot"
          },
          "after": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Snapshot"
              },
              {
                "type": "null"
              }
            ]
          },
          "scoreDelta": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Scoredelta"
          },
          "synergies": {
            "items": {
              "$ref": "#/components/schemas/Synergy"
            },
            "type": "array",
            "title": "Synergies"
          },
          "contributions": {
            "items": {
              "$ref": "#/components/schemas/Contribution"
            },
            "type": "array",
            "title": "Contributions"
          },
          "notice": {
            "type": "string",
            "title": "Notice"
          },
          "finalized": {
            "type": "boolean",
            "title": "Finalized",
            "default": false
          },
          "score": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Score"
          },
          "criticalIndicators": {
            "items": {
              "$ref": "#/components/schemas/CriticalIndicator"
            },
            "type": "array",
            "title": "Criticalindicators"
          },
          "contributionMethod": {
            "type": "string",
            "const": "shapley",
            "title": "Contributionmethod",
            "default": "shapley"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "modelVersion",
          "dataChecksum",
          "decisions",
          "validation",
          "budget",
          "before",
          "after",
          "scoreDelta",
          "synergies",
          "contributions",
          "notice"
        ],
        "title": "Simulation"
      },
...
      "Snapshot": {
        "properties": {
          "score": {
            "type": "number",
            "title": "Score"
          },
          "cityAverage": {
            "type": "number",
            "title": "Cityaverage"
          },
          "weakestDistrictId": {
            "type": "string",
            "enum": [
              "esil",
              "almaty",
              "saryarka",
              "baikonur",
              "nura"
            ],
            "title": "Weakestdistrictid"
          },
          "criticalCount": {
            "type": "integer",
            "title": "Criticalcount"
          },
          "districts": {
            "items": {
              "$ref": "#/components/schemas/District"
            },
            "type": "array",
            "title": "Districts"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "score",
          "cityAverage",
          "weakestDistrictId",
          "criticalCount",
          "districts"
        ],
        "title": "Snapshot"
      },
...
      "Synergy": {
        "properties": {
          "measureIds": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Measureids"
          },
          "districtId": {
            "type": "string",
            "enum": [
              "esil",
              "almaty",
              "saryarka",
              "baikonur",
              "nura"
            ],
            "title": "Districtid"
          },
          "description": {
            "type": "string",
            "title": "Description"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "measureIds",
          "districtId",
          "description"
        ],
        "title": "Synergy"
      },
      "Validation": {
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "valid",
              "invalid"
            ],
            "title": "Status"
          },
          "errors": {
            "items": {
              "$ref": "#/components/schemas/Issue"
            },
            "type": "array",
            "title": "Errors"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "status"
        ],
        "title": "Validation"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}